requirements.txt
-f /usr/share/pip-wheels
alabaster==0.7.6
alembic==0.6.5
aniso8601==0.92
async==0.6.2
Babel==1.3
beautifulsoup4==4.3.2
bleach==1.4.1
blinker==1.3
boto==2.38.0
bottle==0.12.7
Bottleneck==0.8.0
bottlenose==0.6.2
BTrees==4.1.4
cairocffi==0.7.1
CairoSVG==1.0.15
cffi==1.1.2
Chameleon==2.16
chardet==2.3.0
CherryPy==3.7.0
click==5.1
cligj==0.3.0
configobj==5.0.6
CoolProp==5.0.3
coverage==4.0a6
cryptography==1.0.2
cssselect==0.9.1
cvxopt==1.1.6
Cython==0.20.1
dataset==0.5.4
decorator==3.4.0
Django==1.6.6
django-blog-zinnia==0.15.2
django-bootstrap-form==3.1
django-bootstrap3==6.2.2
django-classy-tags==0.6.2
django-cms==3.1.2
django-contrib-comments==1.6.1
django-mongodb-engine==0.6.0
django-mptt==0.7.4
django-sekizai==0.8.2
django-tagging==0.4
django-treebeard==3.0
django-xmlrpc==0.1.5
djangocms-admin-style==0.2.8
djangotoolbox==1.8.0
docker-py==1.4.0
dockerspawner==0.0.1.dev0
docopt==0.6.2
docutils==0.11
EasyProcess==0.1.9
ecdsa==0.13
escapism==0.0.1
falcon==0.1.10
feedgenerator==1.7
feedparser==5.1.3
filebrowser-safe==0.4.0
Fiona==1.5.1
Flask==0.10.1
Flask-Admin==1.0.9
Flask-Assets==0.10
Flask-Babel==0.9
Flask-Bcrypt==0.6.0
Flask-Bootstrap==3.3.4.1
Flask-Email==1.4.4
Flask-HTTPAuth==2.2.1
Flask-Images==2.1.2
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-Moment==0.4.0
Flask-Resize==0.5.2
Flask-RESTful==0.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
future==0.9.0
futures==2.1.6
Genshi==0.7
gensim==0.10.0
geopy==1.10.0
gevent==1.1a1
gitdb==0.6.4
GitPython==1.0.1
gmpy==1.17
google-api-python-client==1.4.1
grappelli-safe==0.4.1
greenlet==0.4.7
grokcore.component==2.6
html5lib==1.0b6
http-parser==0.8.3
httplib2==0.9
idna==2.0
IMAPClient==0.13
infinity==1.3
intervals==0.3.1
ipython==3.2.1
ipython-genutils==0.1.0
isodate==0.5.0
itsdangerous==0.24
jellyfish==0.5.0
Jinja2==2.7.3
jsmin==2.1.4
jsonschema==2.5.1
jupyterhub==0.2.0
Logbook==0.11.2
lxml==3.3.5
Mako==1.0.0
marisa-trie==0.6
Markdown==2.5.2
MarkupSafe==0.23
martian==0.15
matplotlib==1.4.2
Mezzanine==3.1.10
mlpy==3.5.0
mock==1.0.1
mongoengine==0.10.0
mots-vides==2015.5.11
mysql-connector-python==1.2.2
netCDF4==1.1.1
netifaces==0.10.4
networkx==1.8.1
nltk==3.0.0
nose==1.3.3
numexpr==2.4
numpy==1.8.1
oauth2client==1.5.1
oauthlib==1.0.3
openpyxl==1.8.6
pandas==0.14.1
paramiko==1.15.2
Paste==2.0.2
PasteDeploy==1.5.2
pbr==1.8.1
peewee==2.4.4
pelican==3.6.0
pep8==1.5.7
persistent==4.1.1
pexpect==3.3
pilkit==1.1.13
Pillow==2.5.1
pillowcase==2.0.0
prettytable==0.7.2
psutil==2.1.1
psycopg2==2.5.3
ptyprocess==0.5
pudb==2014.1
py-bcrypt==0.4
pyasn1==0.1.8
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyDatalog==0.14.5
pydub==0.9.5
pyenchant==1.6.6
pyfits==3.2.4
pyflakes==0.9.2
pyflakes3k==0.4.3
pygal==1.4.6
pygeoip==0.3.2
Pygments==1.6
pygobject==3.12.0
Pykka==1.2.0
pymc==2.3.2
pymongo==2.8
pyOpenSSL==0.15.1
pyparsing==2.0.3
Pyphen==0.9.1
pyquery==1.2.8
pyramid==1.5.1
pyserial==2.7
pyspotify==2.0.0
PyStemmer==1.3.0
pytesseract==0.1.6
python-amazon-simple-product-api==1.5.0
python-apt===0.9.3.5ubuntu1
python-dateutil==2.2
python-Levenshtein==0.12.0
python-magic==0.4.6
python-mimeparse==0.1.4
python-slugify==0.0.7
pytz==2015.6
PyYAML==3.11
pyzmq==14.3.1
qrcode==5.1
Quandl==2.2
rdflib==4.1.2
redis==2.10.1
repoze.lru==0.6
requests==2.3.0
requests-cache==0.4.4
requests-oauthlib==0.5.0
rpy2==2.5.2
rsa==3.2
ruffus==2.5
scikit-image==0.10.1
scikit-learn==0.15.2
scipy==0.14.0
selenium==2.42.1
setproctitle==1.1.8
Shapely==1.5.9
simplejson==3.7.3
simplepam==0.1.5
simpy==3.0.5
singledispatch==3.4.0.3
six==1.7.2
smmap==0.9.0
snowballstemmer==1.2.0
socketpool==0.5.3
sockjs-tornado==1.0.1
SocksiPy-branch==1.1
South==0.8.4
SPARQLWrapper==1.6.4
speaklater==1.3
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
splinter==0.7.2
SQLAlchemy==1.0.4
SQLAlchemy-Searchable==0.9.3
SQLAlchemy-Utils==0.30.8
stevedore==0.15
stripe==1.17.0
sudospawner==0.0.1
svn==0.3.24
sympy==0.7.5
TA-Lib==0.4.8
tables==3.1.1
Tempita==0.5.3.dev0
terminado==0.5
textblob==0.8.4
tinycss==0.3
tornado==4.2.1
traitlets==4.0.0
transaction==1.4.4
translationstring==1.1
tweepy==3.3.0
twilio==3.6.13
twitter==1.14.3
tzlocal==1.0
unattended-upgrades==0.1
Unidecode==0.4.16
uritemplate==0.6
urllib3==1.10.4
urwid==1.2.1
validators==0.7
venusian==0.9
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
w3lib==1.11.0
Wand==0.3.8
WeasyPrint==0.22
webassets==0.10.1
WebOb==1.4
websocket-client==0.32.0
Werkzeug==0.9.6
wheel==0.26.0
wheezy.caching==0.1.112
wheezy.core==0.1.140
wheezy.html==0.1.145
wheezy.http==0.1.340
wheezy.routing==0.1.153
wheezy.security==0.1.61
wheezy.template==0.1.155
wheezy.validation==0.1.130
wheezy.web==0.1.450
Whoosh==2.7.0
WTForms==2.0.1
WTForms-Components==0.9.7
xlrd==0.9.3
xlwt==1.0.0
z3c.pt==3.0.0a1
zc.buildout==2.4.0
zc.lockfile==1.1.0
ZConfig==3.0.4
zdaemon==4.1.0
ZEO==4.2.0b1
zipline==0.7.0
ZODB==4.2.0
ZODB3==3.11.0
zodbpickle==0.6.0
zope.annotation==4.4.1
zope.app.appsetup==4.0.0a1
zope.app.publication==4.0.0a1.dev0
zope.app.wsgi==4.0.0a4
zope.authentication==4.2.1
zope.browser==2.1.0
zope.browserpage==4.1.0
zope.browserresource==4.1.0
zope.catalog==4.1.0
zope.component==4.2.2
zope.configuration==4.0.3
zope.container==4.1.0
zope.contentprovider==4.0.0
zope.contenttype==4.1.0
zope.datetime==4.1.0
zope.deprecation==4.1.1
zope.dottedname==4.1.0
zope.error==4.2.0
zope.event==4.0.3
zope.exceptions==4.0.7
zope.filerepresentation==4.1.0
zope.formlib==4.3.0
zope.i18n==4.0.1
zope.i18nmessageid==4.0.3
zope.index==4.1.0
zope.interface==4.1.1
zope.intid==4.1.0
zope.keyreference==4.1.0
zope.lifecycleevent==4.1.0
zope.location==4.0.3
zope.login==2.0.0
zope.minmax==2.1.0
zope.pagetemplate==4.2.1
zope.password==4.1.0
zope.principalregistry==4.0.0
zope.processlifetime==2.1.0
zope.proxy==4.1.6
zope.ptresource==4.0.0
zope.publisher==4.2.1
zope.schema==4.4.2
zope.security==4.0.3
zope.securitypolicy==4.0.0
zope.session==4.1.0
zope.site==4.0.0
zope.size==4.1.0
zope.tal==4.1.1
zope.tales==4.1.1
zope.testing==4.2.0
zope.traversing==4.0.0
zope.viewlet==4.0.0
